+++ app/gui/src/project-view/components/CodeMirrorRoot.vue
@@ -21,6 +21,13 @@ defineExpose({ highlightClasses })
   cursor: text;
 }
 
+/* CodeMirror shifts the cursor a bit to be more between letter. But it expects every line having 
+  some padding - otherwise the cursor is clipped off when at the line beginning. We want to set 
+  0 padding in some circumstances, so we remove the offset. */
+.CodeMirrorRoot :deep(.cm-cursor) {
+  margin: 0;
+}
+
 /*
  * Change the looks of codemirror placeholder nodes to look the same as placeholder widgets.
  * This is visible when a `placeholder` property is used in the `useCodeMirror` composable,