+++ app/gui/src/project-view/composables/backend.ts
@@ -58,13 +58,16 @@ type MutationOptions<Method extends BackendMutationMethod> = ToValue<
         Parameters<Backend[Method]>
       >
     >,
-    'mutationFn' | 'mutationKey'
+    'mutationFn'
   > & { invalidate?: boolean }
 >
 
-function backendMutationOptions<Method extends BackendMutationMethod>(
+/**
+ * Create Tanstack Query mutation options for given backend method call.
+ */
+export function backendMutationOptions<Method extends BackendMutationMethod>(
   method: Method,
-  backend: Backend | null,
+  backend: ToValue<Backend | null>,
   options?: MutationOptions<Method>,
 ): UseMutationOptions<
   Awaited<ReturnType<Backend[Method]>> | undefined,
@@ -73,17 +76,20 @@ function backendMutationOptions<Method extends BackendMutationMethod>(
 > {
   return computed(() => {
     const opts = toValue(options)
+    const backendVal = toValue(backend)
     const invalidates =
       opts?.invalidate === false ?
         []
       : (INVALIDATION_MAP[method]?.map((queryMethod) =>
-          queryMethod === INVALIDATE_ALL_QUERIES ? [backend?.type] : [backend?.type, queryMethod],
+          queryMethod === INVALIDATE_ALL_QUERIES ?
+            [backendVal?.type]
+          : [backendVal?.type, queryMethod],
         ) ?? [])
     return {
-      ...backendBaseOptions(backend),
-      mutationKey: [backend?.type, method],
-      mutationFn: (args) => (backend ? (backend[method] as any)(...args) : undefined),
+      ...backendBaseOptions(backendVal),
       ...opts,
+      mutationKey: [backendVal?.type, method, ...(toValue(opts?.mutationKey) ?? [])],
+      mutationFn: (args) => (backendVal ? (backendVal[method] as any)(...args) : undefined),
       meta: {
         invalidates,
         awaitInvalidates: true,